requirements.txt
flask>=3.0.0
newspaper4k>=0.9.2
scikit-learn>=1.4.0
pandas>=2.1.4
numpy>=1.26.2
nltk>=3.8.1
sumy>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.1.0